test(user): add unit test for CreateUserUseCase
Add a unit test case for the CreateUserUseCase class. The test verifies that the use case properly creates a user associated with a tenant. It executes the CreateTenantUseCase to generate a tenant ID and then creates a user using the generated ID. The test ensures the user is created successfully and contains the 'id' property.