Feature/implement multi tenant support (#1)
* chore: add @types/uuid lib

* chore(deps): install @faker-js/faker@^8.2.0
Install @faker-js/faker package version 8.2.0 in the project dependencies.

* feat(tenant): implement Tenant class with UUID generation

Implement the Tenant class with properties 'id', 'name', and 'email'. The 'id' property is generated using the uuid v4 function. Validate input parameters to ensure 'name' and 'email' are provided. Throw an error for invalid params.

* feat(entity): implement User entity

implement the User entity with proprieties 'id', 'name', 'email', 'password', 'tenantId' and 'role'

* feat(useCase): implement CreateTenantUseCase class

Implement the CreateTenantUseCase class to handle the creation of tenants in the system.

* test(createTenant): add unit test for CreateTenantUseCase

Add a unit test case for the CreateTenantUseCase class. The test ensures that the use case properly creates a tenant, including the associated user, and returns an object with the 'id' property.

* feat(useCase): implement CreateUserUseCase class

Implement the CreateUserUseCase class to handle user creation based on the provided input data. The use case checks if a user with the given email already exists. If not, it creates a new user entity and saves it using the UserRepository. The use case returns the created user information.

* test(user): add unit test for CreateUserUseCase

Add a unit test case for the CreateUserUseCase class. The test verifies that the use case properly creates a user associated with a tenant. It executes the CreateTenantUseCase to generate a tenant ID and then creates a user using the generated ID. The test ensures the user is created successfully and contains the 'id' property.

* feat(repository): add TenantRepository interface

Introduce the TenantRepository interface, defining methods for saving, finding by email, updating, and deleting tenant entities. This interface will be implemented by specific repository classes to interact with the persistence layer.

* feat(repository): add UserRepository interface

Introduce the UserRepository interface, specifying methods for saving, deleting, updating, and finding users by email. This interface will be implemented by concrete repository classes to handle user-related database operations.

* feat(repository): implement TenantRepositoryImpl class

Implement the TenantRepositoryImpl class, providing concrete implementations for saving, deleting, updating, and finding tenants by email. This class handles interactions with the in-memory storage and fulfills the methods defined in the TenantRepository interface.

* feat(repository): implement UserRepositoryImpl class

Implement the UserRepositoryImpl class, providing concrete implementations for saving, deleting, updating, and finding users by email. This class manages user-related data interactions with the in-memory storage, fulfilling the methods defined in the UserRepository interface.